refactor(launcher): host activator protocol + one file per terminal/IDE
SessionLauncher.swift was a 200-line monolith mixing a term-program→bundle-id
map, a hard-coded if-ladder dispatch, two AppleScript helpers for iTerm and
Terminal, and the AX+title matcher used by everything else. Adding a host
meant editing 5+ places in one file.

Carve it into:
  - HostActivator protocol (termProgram + bundleID + activate)
  - Shared utilities: AppleScriptRunner, AppActivator, AccessibilityPermission
  - One activator per strategy: ITermActivator (AppleScript by unique id),
    TerminalAppActivator (AppleScript by tty), AXTitleMatchActivator
    (parameterised; reused for every AX-based host)
  - SessionLauncher.swift shrinks to a static registry + jump() + bundleID()

Adding a host is now: append one line to SessionLauncher.activators. If the
strategy is AX+title, no new file at all. If it needs something custom,
one new file under Managers/Launcher/Activators/.

Go side: move termProgramByAppName + termProgramForAppPath out of
osutil_darwin.go into a small hosts_darwin.go with a header comment that
pins the Go-side map to the Swift registry, so the two sides stay in
lockstep when onboarding new hosts.

No public API change; callers in SessionManager.swift and SessionListView.swift
still invoke SessionLauncher.jump(_:). Tests retargeted to the activator
types that own each helper (ITermActivator.iTermUUID, AXTitleMatchActivator
.titleMatchScore/.bestMatchIndex). All 35 tests green.
