+++ core/adapters/inbound/agents/processlifecycle/osutil_darwin.go
@@ -6,7 +6,6 @@ import (
 	"context"
 	"fmt"
 	"os/exec"
-	"path/filepath"
 	"strconv"
 	"strings"
 	"time"
@@ -34,39 +33,6 @@ func readProcessEnv(pid int) (map[string]string, error) {
 // inside VS Code's integrated terminal (claude → zsh → Code Helper → Code);
 // ten gives generous headroom for tmux / SSH nesting.
 const maxAncestry = 10
-
-// termProgramByAppName maps the `.app` bundle name from a process's
-// executable path to the canonical $TERM_PROGRAM string the Swift
-// launcher dispatcher expects. Keys match the directory segment that
-// precedes `.app/Contents/MacOS/` on macOS.
-var termProgramByAppName = map[string]string{
-	"iTerm":              "iTerm.app",
-	"Terminal":           "Apple_Terminal",
-	"Visual Studio Code": "vscode",
-	"Cursor":             "cursor",
-	"Windsurf":           "windsurf",
-	"Ghostty":            "ghostty",
-	"WezTerm":            "WezTerm",
-	"Hyper":              "Hyper",
-	"Warp":               "Warp",
-}
-
-// termProgramForAppPath extracts the host app's canonical TERM_PROGRAM
-// value from an executable path of the form
-// `/Applications/<App>.app/Contents/MacOS/<binary>`. Returns "" for paths
-// without an `.app` segment or when the app isn't one of the hosts we
-// know how to activate.
-func termProgramForAppPath(cmdPath string) string {
-	// Split on ".app/" to find the bundle boundary. The element before the
-	// split contains the app's display name as its final path segment.
-	idx := strings.Index(cmdPath, ".app/")
-	if idx < 0 {
-		return ""
-	}
-	head := cmdPath[:idx]
-	appName := filepath.Base(head)
-	return termProgramByAppName[appName]
-}
 
 // resolveTermProgramFromAncestry walks the parent-process chain of pid and
 // returns the first recognized host app's TERM_PROGRAM string. Returns ""
